+++ src/api.js
@@ -8,20 +8,20 @@
          * Some events have a data property (event.data) which contains data specific to the event.
          *
          * # How to listen to events
-         * Attached to player instance via {@link MTVNPlayer.Player#bind}:
-         *      player.bind("onMetadata",function(event) {
+         * Attached to player instance via {@link MTVNPlayer.Player#on}:
+         *      player.on("metadata",function(event) {
          *             var metadata = event.data;
          *          }
          *      });
          * Passed in as an Object to the constructor {@link MTVNPlayer.Player}:
          *      var player = new MTVNPlayer.Player("video-player",config,{
-         *              onMetadata:function(event) {
+         *              metadata:function(event) {
          *                  var metadata = event.data;
          *              }
          *      });
          * Passed as an Object into {@link MTVNPlayer#createPlayers}
          *      MTVNPlayer.createPlayers("div.MTVNPlayer",config,{
-         *              onMetadata:function(event) {
+         *              metadata:function(event) {
          *                  var metadata = event.data;
          *                  // player that dispatched the event
          *                  var player = event.target;
@@ -30,7 +30,7 @@
          *      });
          * Attached to player from {@link MTVNPlayer#onPlayer}
          *      MTVNPlayer.onPlayer(function(player){
-         *              player.bind("onMetadata",function(event) {
+         *              player.on("metadata",function(event) {
          *                  var metadata = event.data;
          *              }
          *      });
@@ -38,9 +38,9 @@
          */
         MTVNPlayer.Events = {
             /**
-             * @event onMetadata
+             * @event metadata
              * Fired when the metadata changes. event.data is the metadata. Also see {@link MTVNPlayer.Player#currentMetadata}.
-             *      player.bind("onMetadata",function(event) {
+             *      player.on("metadata",function(event) {
              *          // inspect the metadata object to learn more (documentation on metadata is in progress)
              *          console.log("metadata",event.data);
              *
@@ -49,40 +49,56 @@
              *          console.log(event.data === player.currentMetadata); // true
              *      });
              */
-            METADATA: "onMetadata",
+            METADATA: "metadata",
             /**
-             * @event onStateChange
+             * @event stateChange
              * Fired when the play state changes. event.data is the state.
+             * 
+             * You can also listen for a specific state only (v2.5.0).
+             * ```
+             * player.on("stateChange:paused",function(event){
+             *  // callback fires when state equals paused.
+             * });
+             * ```
              */
-            STATE_CHANGE: "onStateChange",
+            STATE_CHANGE: "stateChange",
             /**
-             * @event onMediaStart
+             * @event mediaStart
              * Fired once per playlist item (content + ads/bumpers).
              */
-            MEDIA_START: "onMediaStart",
+            MEDIA_START: "mediaStart",
             /**
-             * @event onMediaEnd
+             * @event mediaEnd
              * Fired when a playlist item ends, this includes its prerolls and postrolls
              */
-            MEDIA_END: "onMediaEnd",
+            MEDIA_END: "mediaEnd",
             /**
-             * @event onPlayheadUpdate
+             * @event playheadUpdate
              * Fired as the playhead moves. event.data is the playhead time.
+             * 
+             * Support for cue points (v2.5.0).
+             * The below snippet fires once when the playhead crosses the 15 second mark.
+             * The playhead time itself may be 15 plus a fraction.
+             * ```
+             * player.one("playheadUpdate:15",function(event){
+             *  // callback
+             * });
+             * ```
              */
-            PLAYHEAD_UPDATE: "onPlayheadUpdate",
+            PLAYHEAD_UPDATE: "playheadUpdate",
             /**
-             * @event onPlaylistComplete
+             * @event playlistComplete
              * Fired at the end of a playlist
              */
-            PLAYLIST_COMPLETE: "onPlaylistComplete",
+            PLAYLIST_COMPLETE: "playlistComplete",
             /**
              * @deprecated 1.5.0 Use {@link MTVNPlayer.Events#onUIStateChange} instead
              * @event onOverlayRectChange
              * Fired when the GUI appears, event.data contains an {Object} {x:0,y:0,width:640,height:320}
              */
-            OVERLAY_RECT_CHANGE: "onOverlayRectChange",
+            OVERLAY_RECT_CHANGE: "overlayRectChange",
             /**
-             * @event onReady
+             * @event ready
              * Fired when the player has loaded and the metadata is available. 
              * You can bind/unbind to events before this fires.
              * You can also invoke most methods before the event, the only exception is
@@ -90,9 +106,9 @@
              * won't be ready until the metadata is ready. Other methods will be queued and 
              * then executed when the player is ready to invoke them.
              */
-            READY: "onReady",
+            READY: "ready",
             /**
-             * @event onUIStateChange
+             * @event uiStateChange
              * Fired when the UI changes its state, ususally due to user interaction, or lack of.
              *
              * event.data will contain information about the state.
@@ -100,27 +116,33 @@
              * If false, the user has remained idle with out interaction for a predetermined amount of time.
              * - data.overlayRect <code>Object</code>: the area that is not obscured by the GUI, a rectangle such as <code>{x:0,y:0,width:640,height:320}</code>
              */
-            UI_STATE_CHANGE: "onUIStateChange",
+            UI_STATE_CHANGE: "uiStateChange",
             /**
-             * @event onIndexChange
+             * @event indexChange
              * Fired when the index of the current playlist item changes, ignoring ads.
              *
              * event.data contains the index
              */
-            INDEX_CHANGE: "onIndexChange",
+            INDEX_CHANGE: "indexChange",
             /**
-             * @event onFullScreenChange
+             * @event fullScreenChange
              * HTML5 only. Fired when the player.isFullScreen property has been changed. 
              * The player may or may not visually be in full screen, it depends on its context.
              * Check {@link MTVNPlayer.Player#isFullScreen} to see if the player is in full screen or not.
              */
-            FULL_SCREEN_CHANGE: "onFullScreenChange",
+            FULL_SCREEN_CHANGE: "fullScreenChange",
             /**
-             * @event onAirplay
+             * @event airplay
              * @private
              * Fired when the airplay button is clicked
              */
-            AIRPLAY: "onAirplay"
+            AIRPLAY: "airplay",
+            /**
+             * @event performance
+             * @private
+             * Fired when performance data has been collected.
+             */
+            PERFORMANCE: "performance"
         };
         /**
          * When a {@link MTVNPlayer.Events#onStateChange} event is fired, the event's data property will be equal to one of these play states. 
@@ -154,8 +176,6 @@
              */
             BUFFERING: "buffering"
         };
-        // swfobject callback
-        MTVNPlayer.onSWFObjectLoaded = null;
         /**
          * @member MTVNPlayer 
          * When using MTVNPlayer.createPlayers this config (see MTVNPlayer.Player.config) object will be used for every player created.
@@ -172,7 +192,7 @@
          * @class MTVNPlayer.Player
          * The player object: use it to hook into events ({@link MTVNPlayer.Events}), call methods, and read properties.
          *      var player = new MTVNPlayer.Player(element/id,config,events);
-         *      player.bind("onMetadata",function(event){console.log("onMetadata",event.data);});
+         *      player.on("metadata",function(event){console.log("metadata",event.data);});
          *      player.pause();
          * @constructor
          * Create a new MTVNPlayer.Player
@@ -189,6 +209,15 @@
                 },
                 document = window.document,
                 Player,
+                fixEventName = function(eventName) {
+                    if (eventName && eventName.indexOf("on") === 0) {
+                        if (eventName === "onUIStateChange") {
+                            return "uiStateChange";
+                        }
+                        return eventName.charAt(2).toLowerCase() + eventName.substr(3);
+                    }
+                    return eventName;
+                },
                 /**
                  * @method checkEventName
                  * @private
@@ -196,18 +225,21 @@
                  * Check if the event exists in our list of events.
                  */
                 checkEventName = function(eventName) {
+                    if (eventName.indexOf(":") !== -1) {
+                        eventName = eventName.split(":")[0];
+                    }
                     var check = function(events) {
-                            for (var event in events) {
-                                if (events.hasOwnProperty(event) && events[event] === eventName) {
-                                    return true; // has event
-                                }
+                        for (var event in events) {
+                            if (events.hasOwnProperty(event) && events[event] === eventName) {
+                                return true; // has event
                             }
-                            return false;
-                        };
+                        }
+                        return false;
+                    };
                     if (check(MTVNPlayer.Events) || check(MTVNPlayer.module("ModuleLoader").Events)) {
                         return;
                     }
-                    throw new Error("MTVNPlayer.Player event:" + eventName + " doesn't exist.");
+                    throwError("event:" + eventName + " doesn't exist.");
                 },
                 /**
                  * @method checkEvents
@@ -217,6 +249,12 @@
                  */
                 checkEvents = function(events) {
                     for (var event in events) {
+                        if (events.hasOwnProperty(event) && event.indexOf("on") === 0) {
+                            events[fixEventName(event)] = events[event];
+                            delete events[event];
+                        }
+                    }
+                    for (event in events) {
                         if (events.hasOwnProperty(event)) {
                             checkEventName(event);
                         }
@@ -300,8 +338,8 @@
              *     MTVNPlayer.onPlayer(function(player){
              *          // player is the player that was just created.
              *          // we can now hook into events.
-             *          player.bind("onPlayheadUpdate",function(event) {
-             *              // do something when "onPlayheadUpdate" fires.
+             *          player.on("playheadUpdate",function(event) {
+             *              // do something when "playheadUpdate" fires.
              *          }
              *
              *          // or look for information about the player.
@@ -345,6 +383,44 @@
             };
             /**
              * @member MTVNPlayer
+             * Returns a player that matches a specific uri
+             * @returns MTVNPlayer.Player
+             */
+            MTVNPlayer.getPlayer = function(uri) {
+                var instances = core.instances,
+                    i = instances.length;
+                for (i; i--;) {
+                    if (instances[i].player.config.uri === uri) {
+                        return instances[i].player;
+                    }
+                }
+                return null;
+            };
+            /**
+             * @member MTVNPlayer
+             * Garbage collection, looks for all {@link MTVNPlayer.Player} that are no longer in the document, 
+             * and removes them from the hash map.
+             */
+            MTVNPlayer.gc = function() {
+                var elementInDocument = function(element) {
+                    while (element.parentNode) {
+                        element = element.parentNode;
+                        if (element == document) {
+                            return true;
+                        }
+                    }
+                    return false;
+                };
+                var instances = core.instances,
+                    i = instances.length;
+                for (i; i--;) {
+                    if (!elementInDocument(instances[i].player.element)) {
+                        instances.splice(i, 1);
+                    }
+                }
+            };
+            /**
+             * @member MTVNPlayer
              * Create players from elements in the page.
              * This should be used if you need to create multiple players that are the same.
              * @param {String} selector default is "div.MTVNPlayer"
@@ -376,10 +452,9 @@
                 if (!selectorQuery) {
                     selectorQuery = "div.MTVNPlayer";
                 }
-                var elements = MTVNPlayer.module("selector").find(selectorQuery),
-                    configModule = MTVNPlayer.module("config");
+                var elements = MTVNPlayer.module("selector").find(selectorQuery);
                 for (var i = 0, len = elements.length; i < len; i++) {
-                    new MTVNPlayer.Player(elements[i], configModule.copyProperties(config || {}, MTVNPlayer.defaultConfig), configModule.copyEvents(events || {}, MTVNPlayer.defaultEvents));
+                    new MTVNPlayer.Player(elements[i], config, events);
                 }
                 return elements.length;
             };
@@ -473,7 +548,7 @@
                  *
                  */
                 this.config = config || {};
-                 /**
+                /**
                  * @property {HTMLElement} isFullScreen
                  * HTML5 only. See {@link MTVNPlayer.Events#onFullScreenChange}
                  */
@@ -491,11 +566,17 @@
                     el = document.getElementById(this.id);
                 }
 
+                if (this.config.performance) {
+                    this.config.performance = {
+                        startTime: (new Date()).getTime()
+                    };
+                }
+
                 // wrap the player element in a container div
                 el.parentNode.insertBefore(containerElement, el);
                 containerElement.appendChild(el);
 
-                this.events = events || {};
+                this.events = MTVNPlayer.module("config").copyEvents(events || {}, MTVNPlayer.defaultEvents);
                 this.isFlash = this.config.isFlash === undefined ? !MTVNPlayer.isHTML5Player : this.config.isFlash;
                 // make sure the events are valid
                 checkEvents(events);
@@ -503,8 +584,8 @@
                 playerModule = MTVNPlayer.module(this.isFlash ? "flash" : "html5");
                 playerModule.initialize();
                 // do more initializing that's across player modules.
-                core.playerInit(this,playerModule);
-                
+                core.playerInit(this, playerModule);
+
                 // check for element before creating
                 if (!el) {
                     if (document.readyState === "complete") {
@@ -615,7 +696,7 @@
                 exitFullScreen: function() {
                     this.message("exitFullScreen");
                 },
-                 /**
+                /**
                  * Show user clip screen.
                  * For flash only (api v2.4.0)
                  */
@@ -624,10 +705,12 @@
                 },
                 /**
                  * Adds an event listener for an event.
+                 * @deprecated use {@link MTVNPlayer.Player#on} instead.
                  * @param {String} eventName an {@link MTVNPlayer.Events}.
                  * @param {Function} callback The function to invoke when the event is fired.
                  */
                 bind: function(eventName, callback) {
+                    eventName = fixEventName(eventName);
                     checkEventName(eventName);
                     var currentEvent = this.events[eventName];
                     if (!currentEvent) {
@@ -641,10 +724,12 @@
                 },
                 /**
                  * Removes an event listener
+                 * @deprecated use {@link MTVNPlayer.Player#off} instead.
                  * @param {String} eventName an MTVNPlayer.Event.
                  * @param {Function} callback The function to that was bound to the event.
                  */
                 unbind: function(eventName, callback) {
+                    eventName = fixEventName(eventName);
                     checkEventName(eventName);
                     var i, currentEvent = this.events[eventName];
                     if (!currentEvent) {
@@ -662,6 +747,7 @@
                 },
                 /**
                  * Adds an event listener for an event that will only fire once and then be removed.
+                 * @deprecated use {@link MTVNPlayer.Player#one} instead.
                  * @param {String} eventName an {@link MTVNPlayer.Events}.
                  * @param {Function} callback The function to invoke when the event is fired.
                  */
@@ -671,9 +757,27 @@
                             ref.unbind(eventName, newCB);
                             callback(event);
                         };
-                    this.bind(eventName, newCB);
+                    this.on(eventName, newCB);
                 }
             };
+            /**
+             * (v2.5.0) Adds an event listener for an event.
+             * @param {String} eventName an {@link MTVNPlayer.Events}.
+             * @param {Function} callback The function to invoke when the event is fired.
+             */
+            Player.prototype.on = Player.prototype.bind;
+            /**
+             * (v2.5.0) Removes an event listener
+             * @param {String} eventName an MTVNPlayer.Event.
+             * @param {Function} callback The function to that was bound to the event.
+             */
+            Player.prototype.off = Player.prototype.unbind;
+            /**
+             * (v2.5.0) Adds an event listener for an event that will only fire once and then be removed.
+             * @param {String} eventName an {@link MTVNPlayer.Events}.
+             * @param {Function} callback The function to invoke when the event is fired.
+             */
+            Player.prototype.one = Player.prototype.once;
             return Player;
         }(window));
         /**