+++ src/util/reporting.js
@@ -1,7 +1,9 @@
 /**
-* Trying something new here. A way to keep the API clean for utility methods specific to things like reporting.
-* These modules are on a player, as opposed to the modules on MTVNPlayer.
-*/
+ * @private
+ * @ignore
+ * Trying something new here. A way to keep the API clean for utility methods specific to things like reporting.
+ * These modules are on a player, as opposed to the modules on MTVNPlayer.
+ */
 (function(MTVNPlayer) {
     "use strict";
     MTVNPlayer.onPlayer(function(player) {